requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-telegram-bot==22.0
rsa==4.9.1
sniffio==1.3.1
Telethon==1.40.0
typing_extensions==4.13.2
tzlocal==5.3.1
